role permission

+++ routes/web.php
@@ -7,6 +7,7 @@ use GuzzleHttp\Middleware;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\Auth\AuthController;
 use App\Http\Middleware\NoCache;
+use Illuminate\Support\Facades\Request;
 
 Route::get('/user', function (Request $request) {
     return $request->user();
@@ -32,9 +33,9 @@ Route::prefix('admin')->controller(AuthController::class)->group(function(){
 //getting view With Auth Verifi
 
 Route::prefix('admin')->controller(HomeController::class)->middleware(['auth','verified'])->group(function(){
-    Route::get('admin-management','adminall')->name('admin.all');
+    Route::get('admin-management','adminall')->name('admin.all')->middleware('permission:view');
     Route::get('dashboard','dashboard');
-    Route::get('department-management','adddepartment')->name('add.dep');
+    Route::get('department-management','adddepartment')->name('add.dep')->middleware('permission:view');
      Route::get('viewdep','viewdep')->name('view.dep');
 });
 
@@ -47,10 +48,10 @@ Route::get('/', function () {
 });
 
 Route::prefix('admin')->middleware(['auth','verified'])->controller(DepartmentController::class)->group(function(){
-    Route::post('new-department','newDepartment')->name('new.department');
-    Route::get('departments/{id}','getDepartment');
-    Route::put('departments/{id}','updateDepartmnet');
-    Route::delete('departments/{id}','destroy');
+    Route::post('new-department','newDepartment')->name('new.department')->middleware('permission:create');
+    Route::get('departments/{id}','getDepartment')->middleware('permission:view');
+    Route::put('departments/{id}','updateDepartmnet')->middleware('permission:edit');
+    Route::delete('departments/{id}','destroy')->middleware('permission:delete');
     Route::post('new-admin','newAdmin')->name('new.admin');
     Route::get('admin/{id}','getAdmin');
     Route::put('adminup/{id}','updateAdmin');
